♻️ Refactor(#2): 내부타입 이름 개선
